[refactor] fold dynamicemb e2e tests + rename DP proposer test + bump copyright
(5) Move `PlanUtilDynamicEmbE2ETest` from the standalone
`tzrec/utils/plan_util_dynamicemb_e2e_test.py` into `plan_util_test.py`
where the rest of the plan_util tests already live. Delete the standalone
file.

(6) Rename `DynamicProgrammingProposer2DTest` -> `DynamicProgrammingProposerTest`.
The proposer class itself stays `DynamicProgrammingProposer`.

(7) Bump copyright `2024 -> 2026` in the new file
`tzrec/utils/dynamicemb_util_test.py`. The header in
plan_util_test.py stays at 2024 since it was a pre-existing file we only
extended.
